moved ssn validation logic to shared dir and updated 1010d usage
