+++ src/applications/ivc-champva/shared/validations.js
@@ -82,3 +82,53 @@ export const certifierAddressCleanValidation = (errors, page, formData) => {
 export const applicantAddressCleanValidation = (errors, page, formData) => {
   return validAddressCharsOnly(errors, page, formData);
 };
+
+export function noDash(str) {
+  return str?.replace(/-/g, '');
+}
+
+/**
+ * Validates the sponsor's SSN does not match any others in the form.
+ * @param {Object} errors - The errors object for the current page
+ * @param {Object} page - The current page data
+ */
+export const validateSponsorSsnIsUnique = (errors, page) => {
+  const keyname = page.ssn ? 'ssn' : 'sponsorSsn'; // accomodate 10-10D and 10-10d-extended
+  const sponsorSSN = page[keyname];
+  const match = page?.applicants?.find(
+    el => noDash(el?.applicantSSN) === noDash(sponsorSSN),
+  );
+
+  if (match) {
+    errors?.[keyname]?.addError(
+      'This Social Security number is in use elsewhere in the form. SSNs must be unique.',
+    );
+  }
+};
+
+/**
+ * Validates that an applicant's SSN does not match any others in the form.
+ * Relies on `view:` properties set from the applicant SSN page since full
+ * form data isn't available to list loop v1 pages outside a custom page.
+ * @param {Object} errors - The errors object for the current page
+ * @param {Object} page - The current page data
+ */
+export const validateApplicantSsnIsUnique = (errors, page) => {
+  const idx = page?.['view:pagePerItemIndex'];
+  if (idx === undefined) return; // We can't process without this
+  const sponsorMatch =
+    noDash(page?.applicantSSN) === noDash(page?.['view:sponsorSSN']);
+
+  let applicants = page?.['view:applicantSSNArray'];
+  if (applicants === undefined) return; // We can't process without this
+  applicants = [...applicants.slice(0, idx), ...applicants.slice(idx + 1)];
+  const applicantMatch = applicants?.some(
+    app => noDash(app) === noDash(page?.applicantSSN),
+  );
+
+  if (sponsorMatch || applicantMatch) {
+    errors.applicantSSN.addError(
+      'This Social Security number is in use elsewhere in the form. SSNs must be unique.',
+    );
+  }
+};
